requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
pydantic==2.8.2
email-validator==2.2.0
pydantic-settings==2.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.4.0
redis==5.0.8
requests==2.32.3
# Ticketing
reportlab==4.2.2
qrcode==7.4.2
pillow==10.4.0

# Optional integrations
google-cloud-storage==2.18.2
